+++ src/main/java/insane96mcp/enhancedai/modules/shulker/ShulkerBullets.java
@@ -1,5 +1,6 @@
 package insane96mcp.enhancedai.modules.shulker;
 
+import insane96mcp.enhancedai.data.EAIData;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.LoadFeature;
@@ -8,23 +9,48 @@ import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.Difficulty;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.monster.Shulker;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@LoadFeature(module = Modules.Ids.SHULKER)
+import javax.annotation.Nullable;
+
+@LoadFeature(module = Modules.Ids.SHULKER, description = "Shulker bullets' levitation duration and amplifier based off owner's data")
 public class ShulkerBullets extends Feature {
     @Config(min = 1, max = 600)
     public static Difficulty levitationDuration = new Difficulty(100, 100, 160);
     @Config(min = 0, max = 127, description = "Note that 0 = I, 1 = II, and so on")
     public static Difficulty levitationAmplifier = new Difficulty(5, 5, 5);
 
-    public ShulkerBullets(Module module, boolean enabledByDefault, boolean canBeDisabled) {
-        super(module, enabledByDefault, canBeDisabled);
+    public static EAIData<Integer> LEVITATION_DURATION;
+    public static EAIData<Integer> LEVITATION_AMPLIFIER;
+
+    public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+        super.init(module, enabledByDefault, canBeDisabled);
+        LEVITATION_DURATION = EAIData.ofInt(this.createDataKey("levitation_duration"));
+        LEVITATION_AMPLIFIER = EAIData.ofInt(this.createDataKey("levitation_amplifier"));
     }
 
-    public static MobEffectInstance getLevitationInstance(Level level, MobEffectInstance originalValue) {
+    public static MobEffectInstance getLevitationInstance(Level level, MobEffectInstance originalValue, @Nullable Entity owner) {
         if (!isEnabled(ShulkerBullets.class))
             return originalValue;
 
-        return new MobEffectInstance(MobEffects.LEVITATION, (int) levitationDuration.getByDifficulty(level), (int) levitationAmplifier.getByDifficulty(level));
+        if (!(owner instanceof Mob mob))
+            return new MobEffectInstance(MobEffects.LEVITATION, (int) levitationDuration.getByDifficulty(level), (int) levitationAmplifier.getByDifficulty(level));
+        else return new MobEffectInstance(MobEffects.LEVITATION, LEVITATION_DURATION.get(mob), LEVITATION_AMPLIFIER.get(mob));
+    }
+
+    @SubscribeEvent
+    public void onSpawn(EntityJoinLevelEvent event) {
+        if (!this.isEnabled()
+                || event.getLevel().isClientSide
+                || !(event.getEntity() instanceof Shulker shulker))
+            return;
+
+        LEVITATION_DURATION.applyIfAbsent(shulker, (int) levitationDuration.getByDifficulty(event.getLevel()));
+        LEVITATION_AMPLIFIER.applyIfAbsent(shulker, (int) levitationAmplifier.getByDifficulty(event.getLevel()));
     }
 }